Added providers + relocated snakeyaml
